refactor(db): Split collection.rs into directory module
Split collection.rs (1,049 lines) into 5 focused files:
- mod.rs: struct Collection, constants, collection_short_id(), key helpers
- index_ops.rs: create/update/delete_with_indexes(), is_deleted()
- crud.rs: create/get/update/delete/exists/save/iterate/get_all (DbTxn path)
- crud_datastore.rs: *_with_datastore() variants (NamespaceView path)
- validation.rs: validate_document() + type compatibility functions

Key helper methods (doc_key, deleted_key, version_key, etc.) changed from
private to pub(crate) so cross-file impl Collection blocks can access them.
